Remove dependence on nokiaui and checking for supported Nokia platforms (now will work on phones that don't support nokiaui), removed unused CommandListener in GameRuntime
nokiaui was used for drawing translucent polygons (for bg of buttons and pause menu and for water) and to control backlight. I've written my own realization of transparency, it is now used for drawing UI, but not for drawing water because this workaround is too slow to give reasonable FPS. nokiaui is still being used if the phone supports it.